rename analysis package to dynamicanalysis
