+++ internal/dynamicanalysis/output_test.go
@@ -1,4 +1,4 @@
-package analysis
+package dynamicanalysis
 
 import (
 	"bytes"
